refactor: move hydration in a different file
