+++ packages/@lwc/engine-core/src/framework/api.ts
@@ -38,16 +38,22 @@ import {
     VNodeType,
 } from '../3rdparty/snabbdom/types';
 import { LightningElementConstructor } from './base-lightning-element';
-import {
-    markAsDynamicChildren,
-    // TextHook,
-    // CommentHook,
-    // ElementHook,
-    // CustomElementHook,
-} from './hooks';
 import { isComponentConstructor } from './def';
 
 const SymbolIterator: typeof Symbol.iterator = Symbol.iterator;
+
+// Using a WeakMap instead of a WeakSet because this one works in IE11 :(
+const FromIteration: WeakMap<VNodes, 1> = new WeakMap();
+
+// dynamic children means it was generated by an iteration
+// in a template, and will require a more complex diffing algo.
+function markAsDynamicChildren(children: VNodes) {
+    FromIteration.set(children, 1);
+}
+
+export function hasDynamicChildren(children: VNodes): boolean {
+    return FromIteration.has(children);
+}
 
 // [h]tml node
 function h(sel: string, data: VElementData, children: VNodes): VElement {
@@ -102,7 +108,6 @@ function h(sel: string, data: VElementData, children: VNodes): VElement {
         text,
         elm,
         key,
-        // hook: ElementHook,
         owner: vmBeingRendered,
     };
 }
@@ -218,8 +223,6 @@ function c(
         text,
         elm,
         key,
-
-        // hook: CustomElementHook,
         ctor: Ctor,
         owner: vmBeingRendered,
         mode: 'open', // TODO [#1294]: this should be defined in Ctor
@@ -353,8 +356,6 @@ function t(text: string): VText {
         text,
         elm,
         key,
-
-        // hook: TextHook,
         owner: getVMBeingRendered()!,
     };
 }
@@ -371,8 +372,6 @@ function co(text: string): VComment {
         text,
         elm,
         key,
-
-        // hook: CommentHook,
         owner: getVMBeingRendered()!,
     };
 }
