+++ packages/@lwc/engine-core/src/framework/vm.ts
@@ -33,7 +33,8 @@ import {
     logGlobalOperationEnd,
     logGlobalOperationStart,
 } from './profiler';
-import { hydrateChildrenHook, patchChildren } from './hooks';
+import { patchChildren } from './rendering';
+import { hydrateChildren } from './hydration';
 import { ReactiveObserver } from './mutation-tracker';
 import { connectWireAdapters, disconnectWireAdapters, installWireAdapters } from './wiring';
 import { AccessorReactiveObserver } from './decorators/api';
@@ -72,6 +73,10 @@ export const enum ShadowMode {
 export const enum ShadowSupportMode {
     Any = 'any',
     Default = 'reset',
+}
+
+export const enum LWCDOMMode {
+    manual = 'manual',
 }
 
 export interface Context {
@@ -429,7 +434,7 @@ function hydrate(vm: VM) {
 
         const vmChildren =
             vm.renderMode === RenderMode.Light ? vm.elm.childNodes : vm.elm.shadowRoot.childNodes;
-        hydrateChildrenHook(vmChildren, children, vm);
+        hydrateChildren(vmChildren, children, vm);
 
         runRenderedCallback(vm);
     }
